making a pages, sidebar component
