+++ package.json
@@ -13,6 +13,7 @@
     "bootstrap": "^5.3.2",
     "jquery": "^3.7.1",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.9.1",
     "react-dom": "^18.2.0",
     "react-icons": "^4.12.0"
   },
