+++ src/assets/TodayCard.css
@@ -1,34 +1,41 @@
 .add-todo-input,
 .edit-todo-input {
-    outline: none;
+  outline: none;
 }
 
 .add-todo-input:focus,
 .edit-todo-input:focus {
-    border: none !important;
-    box-shadow: none !important;
+  border: none !important;
+  box-shadow: none !important;
 }
 
 .view-opt-label,
 .date-label {
-    font-size: 0.8rem;
+  font-size: 0.8rem;
 }
 
 .edit-todo-input {
-    font-size: 1.7rem !important;
+  font-size: 1.7rem !important;
 }
 
 .todo-actions {
-    visibility: hidden !important;
+  visibility: hidden !important;
 }
 
 .todo-item:hover .todo-actions {
-    visibility: visible !important;
+  visibility: visible !important;
 }
 
 .todo-item.editing .todo-actions .edit-icon {
-    display: none !important;
+  display: none !important;
 }
-.col-md-9{
-    justify-content: right;
+.col-md-9 {
+  justify-content: right;
+}
+
+.todoapp-container {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: space-between;
 }
