+++ src/assets/style.css
@@ -47,7 +47,7 @@ body {
     line-height: 1.5;
     color: #212529;
     text-align: left;
-    background-color: #fff;
+    background-color: #fff
 }
 
 [tabindex="-1"]:focus {
@@ -10809,8 +10809,7 @@ h1,h2,h3,h4,h5,.h1,.h2,.h3,.h4,.h5 {
 }
 
 .wrapper {
-    width: 100%;
-    background-color: #FFD9B5;
+    width: 100%
 }
 
 #sidebar {
