+++ src/components/todoapp1.jsx
@@ -31,11 +31,12 @@ const TodoApp1 = () => {
     updatedTasks[index] = newValue;
     setTasks(updatedTasks);
   };
-
+  <link rel="stylesheet" href="https://fonts.googleapis.com/css?family=Sofia&effect=fire"></link>
   return (
+    
     <div className="container m-5 p-2 rounded mx-auto bg-light shadow">
       {/* ... (unchanged code) */}
-      <h3>Reward</h3>
+      <h6>Reward</h6>
       <div className="row m-1 p-3">
         <div className="col col-11 mx-auto">
           <div className="row bg-white rounded shadow-sm p-2 add-todo-wrapper align-items-center justify-content-center">
